move TestRow to correct class
