+++ test/publish/test_data_quality.py
@@ -9,7 +9,7 @@ from sqlalchemy import select
 
 from rialto_airflow.publish import data_quality
 from rialto_airflow.database import Publication, Author, Funder
-from test.test_utils import TestRow
+from test.utils import TestRow
 
 
 @pytest.fixture
