+++ test/test_utils.py
@@ -1,6 +1,5 @@
 import csv
 from pathlib import Path
-from dataclasses import dataclass
 
 import pytest
 
@@ -72,16 +71,3 @@ def test_normalize_orcid():
         utils.normalize_orcid(" HTTPS://ORCID.org/0000-0002-7262-6251 ")
         == "0000-0002-7262-6251"
     )
-
-
-@dataclass
-class TestRow:
-    """
-    An object that simulates a Publication database row.
-    """
-
-    dim_json: dict | None = None
-    openalex_json: dict | None = None
-    sulpub_json: dict | None = None
-    wos_json: dict | None = None
-    pubmed_json: dict | None = None
