+++ test/utils.py
@@ -1,4 +1,5 @@
 import json
+from dataclasses import dataclass
 
 
 def num_jsonl_objects(jsonl_path):
@@ -16,3 +17,16 @@ def load_jsonl_file(path):
         for line in f:
             result.append(json.loads(line))
     return result
+
+
+@dataclass
+class TestRow:
+    """
+    An object that simulates a Publication database row.
+    """
+
+    dim_json: dict | None = None
+    openalex_json: dict | None = None
+    sulpub_json: dict | None = None
+    wos_json: dict | None = None
+    pubmed_json: dict | None = None
